moved code to setup_file and added environment variables
