+++ .env
@@ -1,0 +1,3 @@
+XML_PATH=data/xml
+JSON_FILEPATH=data/parsed_data.json
+DATABASE_FILEPATH=data/data.db
